Replace sql with text files + spclient refactoring

+++ src/stream.rs
@@ -3,11 +3,12 @@ use std::sync::Mutex;
 use flutter_rust_bridge::StreamSink;
 use lazy_static::lazy_static;
 
-use crate::constants::{LogEntry, ScanProgress};
+use crate::constants::LogEntry;
+use crate::spclient::ScanProgress;
 
 lazy_static! {
     static ref LOG_STREAM_SINK: Mutex<Option<StreamSink<LogEntry>>> = Mutex::new(None);
-    static ref AMOUNT_STREAM_SINK: Mutex<Option<StreamSink<u32>>> = Mutex::new(None);
+    static ref AMOUNT_STREAM_SINK: Mutex<Option<StreamSink<u64>>> = Mutex::new(None);
     static ref SCAN_STREAM_SINK: Mutex<Option<StreamSink<ScanProgress>>> = Mutex::new(None);
 }
 
@@ -16,7 +17,7 @@ pub fn create_log_stream(s: StreamSink<LogEntry>) {
     *stream_sink = Some(s);
 }
 
-pub fn create_amount_stream(s: StreamSink<u32>) {
+pub fn create_amount_stream(s: StreamSink<u64>) {
     let mut stream_sink = AMOUNT_STREAM_SINK.lock().unwrap();
     *stream_sink = Some(s);
 }
@@ -35,7 +36,7 @@ pub(crate) fn loginfo(text: &str) {
     }
 }
 
-pub(crate) fn send_amount_update(amount: u32) {
+pub(crate) fn send_amount_update(amount: u64) {
     let stream_sink = AMOUNT_STREAM_SINK.lock().unwrap();
     if let Some(stream_sink) = stream_sink.as_ref().clone() {
         stream_sink.add(amount);
